[COVERITY] drm-intel: fix Resource leak and integer overflow (3 findings)
Resource leak in drmModeGetProperty (xf86drmMode.c:698): r->values was
assigned unconditionally then overwritten in the DRM_MODE_PROP_BLOB branch.

Resource leak in drmSLCreate (xf86drmSL.c:~110): if SLCreateEntry returns
NULL, list was returned with a NULL head. Free list and return NULL instead.

Overflowed integer argument in drm_intel_decode (intel_decode.c:3967):
decode_3d() returns signed int; cast to unsigned int before adding to index.

Tracked-on: NIACP3-1264

+++ intel/intel_decode.c
@@ -3964,7 +3964,7 @@ drm_intel_decode(struct drm_intel_decode *ctx)
 				index +=
 				    decode_3d_965(ctx);
 			} else if (IS_GEN3(devid)) {
-				index += decode_3d(ctx);
+				index += (unsigned int)decode_3d(ctx);
 			} else {
 				index +=
 				    decode_3d_i830(ctx);
